requirements.txt
# Core dependencies
asyncio
aiohttp>=3.8.0
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Database
sqlalchemy>=2.0.0
alembic>=1.8.0
psycopg2-binary>=2.9.0  # PostgreSQL support

# Media processing
ffmpeg-python>=0.2.0

# Configuration and validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyYAML>=6.0

# Logging
loguru>=0.7.0

# CLI
click>=8.1.0
rich>=13.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.64.0

# Development dependencies
pytest>=7.2.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0

# Optional: GUI (if needed later)
# streamlit>=1.28.0